Cleanup and add features for exception testing
