+++ .vscode/launch.json
@@ -2,21 +2,16 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "preLaunchTask": "${input:profile}",
+      "preLaunchTask": "probe-rs-debugger-test",
       "type": "probe-rs-debug",
       "request": "launch",
-      "name": "probe-rs-debugger",
-      "chip": "${fileBasenameNoExtension}",
-      // "cwd": "${workspaceFolder}",
-      
-      // "runtimeExecutable": "probe-rs", // OPTIONAL: If you want to use a specific version/install location of probe-rs.
-      // "runtimeArgs": ["debug"], // OPTIONAL: Add additional arguments to the probe-rs command.
-      // "probe": "${config:vid-pid}", // OPTIONAL: This can be omitted if there is only one probe connected.
+      "name": "probe-rs-debugger-test",
+      "chip": "${config:probe-rs-chip}",
       "coreConfigs": [
         {
           "coreIndex": 0,
-          "programBinary": "./target/${fileDirnameBasename}/${input:profile}/${fileBasenameNoExtension}",
-          "svdFile": "./svd_files/${fileBasenameNoExtension}.svd",
+          "programBinary": "./target/${config:rust-analyzer.cargo.target}/${config:probe-rs-profile}/${config:probe-rs-chip}",
+          "svdFile": "./svd_files/${config:probe-rs-chip}.svd",
           "rttEnabled": true,
           "rttChannelFormats": [
             {
@@ -36,35 +31,12 @@
         "haltAfterReset": false
       },
       // "server": "127.0.0.1:50001", // OPTIONAL: For connecting to an existing probe-rs-debugger process.
-      // "connectUnderReset": true,
-      "consoleLogLevel": "Debug",
-      "env": {
-      "RUST_LOG": "debug",
-      "DEFMT_LOG": "debug"
-      },
-      "speed": 1000 // CONFIGURE: Optionally update USB speed
+      // "consoleLogLevel": "Debug",
+      // "env": {
+      // "RUST_LOG": "debug",
+      // "DEFMT_LOG": "debug"
+      // },
+      "speed": 4000 // CONFIGURE: Optionally update USB speed
     }
-  ],
-  "inputs": [
-    {
-      "id": "profile",
-      "type": "pickString",
-      "description": "Please select a profile for level of optimization required.",
-      "options": [
-        {
-          "label": "Debug Info + LTO=off + opt-level = 0",
-          "value": "debug-no-opt"
-        },
-        {
-          "label": "Debug Info + LTO=thin + opt-level = 1",
-          "value": "debug-lto-thin-opt-1"
-        },
-        {
-          "label": "Debug Info + LTO=fat + opt-level = 2",
-          "value": "debug-lto-fat-opt-2"
-        }
-      ],
-      "default": "debug-no-opt"
-    }
-  ]
+    ]
 }
