--- probe-rs-debugger-test.code-workspace
@@ -1,17 +1,0 @@
-{
-  "folders": [
-    {
-      "path": "."
-    }
-  ],
-  "settings": {
-    // The following are convenience settings for code editing.
-    "rust-analyzer.checkOnSave": true,
-    "rust-analyzer.check.overrideCommand": [
-      "cargo",
-      "clippy",
-      "--message-format",
-      "json-diagnostic-rendered-ansi"
-    ]
-  }
-}
