+++ riscv-probe-rs-debugger-test.code-workspace
@@ -1,0 +1,31 @@
+{
+  "folders": [
+    {
+      "path": "."
+    }
+  ],
+  "settings": {
+    // The following are convenience settings for code editing.
+    "rust-analyzer.checkOnSave": true,
+
+    // All entries below must be configured to match the target device and required features.
+    "probe-rs-chip": "esp32c3", // CONFIGURE: Replace with probe-rs supported CHIP name
+    "probe-rs-profile": "debug-no-opt",
+    // PLEASE NOTE: For some reason, these have to be duplicated in the "rust-analyzer.check.overrideCommand" setting.
+    "rust-analyzer.cargo.target": "riscv32imc-unknown-none-elf", // CONFIGURE: Replace target identity
+    "rust-analyzer.cargo.features": ["esp32c3"], // CONFIGURE: Replace with probe-rs supported CHIP name
+    "rust-analyzer.cargo.extraArgs": ["--bin", "esp32c3"], // CONFIGURE: Replace with probe-rs supported CHIP name
+    "rust-analyzer.check.overrideCommand": [
+      "cargo",
+      "clippy",
+      "--message-format",
+      "json-diagnostic-rendered-ansi",
+      "--features",
+      "esp32c3", // CONFIGURE: Replace with probe-rs supported CHIP name
+      "--bin",
+      "esp32c3", // CONFIGURE: Replace with probe-rs supported CHIP name
+      "--target",
+      "riscv32imc-unknown-none-elf" // CONFIGURE: Replace target identity
+    ]
+  }
+}
