--- src/thumbv7em-none-eabihf/STM32H745ZITx.rs
@@ -1,84 +1,0 @@
-#![no_std]
-#![no_main]
-// We want our binary names the same as the chipname.
-#![allow(non_snake_case)]
-// Because this is not code that doesn't need to serve a purpose other than testing the debugger
-#![allow(unused_variables)]
-#![allow(dead_code)]
-#![allow(unused_assignments)]
-#![feature(core_intrinsics)]
-
-// This section contains the declarations that is common to all the tests, and is not specific to any one chip/board.
-use common_testing_code::*;
-use rtt_target::rprintln;
-
-// Board/Chip specific code.
-use cortex_m::peripheral::syst::SystClkSource;
-use cortex_m_rt::{entry, exception, ExceptionFrame};
-
-#[exception]
-fn SysTick() {
-    static mut COUNT: u32 = 0;
-
-    *COUNT += 1;
-
-    if *COUNT == 5 {
-        // If `UsageFault` is enabled, we disable that first, since otherwise `udf` will cause that
-        // exception instead of `HardFault`.
-        const SHCSR: *mut u32 = 0xE000ED24usize as _;
-        const USGFAULTENA: usize = 18;
-        unsafe {
-            let mut shcsr = core::ptr::read_volatile(SHCSR);
-            shcsr &= !(1 << USGFAULTENA);
-            core::ptr::write_volatile(SHCSR, shcsr);
-        }
-        // Cause a UDF, by reading executing an invalid instruction.
-        core::intrinsics::abort()
-    }
-}
-
-#[exception]
-unsafe fn HardFault(ef: &ExceptionFrame) -> ! {
-    rprintln!("HardFault at {:#?}", ef);
-    unsafe {
-        core::arch::asm!("bkpt");
-    }
-    loop {}
-}
-
-#[panic_handler]
-fn panic(info: &core::panic::PanicInfo) -> ! {
-    rprintln!("Panic: {:?}", info);
-    loop {
-        unsafe {
-            core::arch::asm!("bkpt");
-        }
-        rprintln!("In a panic loop, stepped past the breakpoint");
-    }
-}
-
-#[entry]
-fn main() -> ! {
-    // Board/Chip specific code.
-    let p = cortex_m::Peripherals::take().unwrap();
-    let mut syst = p.SYST;
-
-    // configures the system timer to trigger a SysTick exception every second
-    syst.set_clock_source(SystClkSource::Core);
-    // this is configured for the STM32H745 which has a default CPU clock of 400 MHz
-    syst.set_reload(400_000_000);
-    syst.clear_current();
-    syst.enable_counter();
-    syst.enable_interrupt();
-
-    // Common testing code.
-    let (mut loop_counter, mut binary_rtt_channel) = setup_data_types();
-
-    loop {
-        // Common testing code.
-        shared_loop_processing(&mut binary_rtt_channel, &mut loop_counter);
-
-        // Board/Chip specific code.
-        cortex_m::asm::delay(1_000_001);
-    }
-}
